+++ apps/recnet-api/prisma/migrations/20241124232108_add_slack_oauth/down.sql
@@ -1,0 +1,6 @@
+-- AlterTable
+ALTER TABLE "User" DROP COLUMN "slackAccessToken",
+DROP COLUMN "slackUserId",
+DROP COLUMN "slackWorkspaceName",
+ADD COLUMN     "slackEmail" VARCHAR(128);
+