+++ apps/recnet-api/src/config/env.schema.ts
@@ -25,6 +25,11 @@ export const EnvSchema = z.object({
   SMTP_PASS: z.string(),
   // slack config
   SLACK_TOKEN: z.string().optional(),
+  SLACK_CLIENT_ID: z.string(),
+  SLACK_CLIENT_SECRET: z.string(),
+  SLACK_TOKEN_ENCRYPTION_KEY: z
+    .string()
+    .transform((val) => Buffer.from(val, "base64")),
 });
 
 export const parseEnv = (env: Record<string, string | undefined>) => {